Test the debugger rendez-vous fallback via a proc-maps failspot
We move the tests in their own process to make sure we get an isolated
failspot instance.
