+++ src/lib.rs
@@ -14,6 +14,7 @@ cfg_if::cfg_if! {
                 ThreadName,
                 SuspendThreads,
                 CpuInfoFileOpen,
+                EnumerateMappingsFromProc,
             }
         }
